UseBridgeIPasGW and dont require subnet
Add a new config parameter: useBridgeIPasGW. When this is set and IsGW
is set, instead of using the gateway returned by the IPAM plugin or
guessing it based off of the subnet of the IP returned by the IPAM
plugin, use the bridge's existing IP.

Don't make subnet required. It is used to add an IP to the bridge, but
in some usecases, we don't want to supply a subnet and we expect the
bridge to already have an IP address. So, by making subnet optional,
the plugin will not add an IP to the bridge.